Fix in PixelDropout
