+++ albumentations/augmentations/dropout/__init__.py
@@ -2,4 +2,5 @@ from .channel_dropout import *
 from .coarse_dropout import *
 from .grid_dropout import *
 from .mask_dropout import *
+from .transforms import *
 from .xy_masking import *
